move tests into different files
